refactor utils and add formatted_time
