+++ src/autogluon/bench/datasets/object_detection_dataset.py
@@ -1,10 +1,10 @@
 import abc
 import os
 
+from autogluon.bench.utils.dataset_utils import get_data_home_dir, get_repo_url
 from autogluon.common.loaders import load_zip
 
 from .constants import _OBJECT_DETECTION
-from .utils import get_data_home_dir, get_repo_url
 
 # Add dataset class names here
 __all__ = ["TinyMotorbike", "Clipart"]
